requirements.txt
streamlit>=1.50.0
python-dotenv
pandas
colorama 
PyYAML
pymongo


